Moved more common stuff into common Fixed some (unit) tests
