+++ CfpExchange.Common/Constants/Constants.cs
@@ -5,5 +5,10 @@
         public const string QUEUE_IMAGES = "eventimages";
 
         public const string QUEUE_TWITTER = "newcfptweet";
+
+        public static string NoEventImageUrl = $"{WebsiteRootUrl}/images/noimage.svg";
+
+        public const string WebsiteRootUrl = "https://cfp.exchange";
+
     }
 }
