+++ CfpExchange.Common/Models/TimezoneInfo.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using Newtonsoft.Json;
 
-namespace CfpExchange.Models
+namespace CfpExchange.Common.Models
 {
 	public class TimezoneInfo
 	{
