--- CfpExchange/Helpers/Constants.cs
@@ -1,8 +1,0 @@
-﻿namespace CfpExchange.Helpers
-{
-	public static class Constants
-	{
-		public static string NoEventImageUrl = $"{WebsiteRootUrl}/images/noimage.svg";
-		public const string WebsiteRootUrl = "https://cfp.exchange";
-	}
-}
