+++ CfpExchange/Models/Cfp.cs
@@ -2,7 +2,7 @@
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 using CfpExchange.Common.Enums;
-using CfpExchange.Helpers;
+using CfpExchange.Common;
 
 namespace CfpExchange.Models
 {
