+++ CfpExchange/Startup.cs
@@ -11,15 +11,13 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
+using CfpExchange.Common;
 using CfpExchange.Data;
-using CfpExchange.Helpers;
 using CfpExchange.Middleware;
 using CfpExchange.Models;
 using CfpExchange.Services;
 using CfpExchange.Common.Services.Interfaces;
 using CfpExchange.Common.Services;
-using HtmlAgilityPack;
-using System.Collections;
 
 namespace CfpExchange
 {
