+++ CfpExchange/ViewModels/ErrorViewModel.cs
@@ -1,0 +1,9 @@
+﻿namespace CfpExchange.ViewModels
+{
+    public class ErrorViewModel
+	{
+		public string RequestId { get; set; }
+
+		public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
+	}
+}
