+++ CfpExchange/Views/Cfp/Submit.cshtml
@@ -41,7 +41,7 @@
 
                         <div class="form-group" id="eventUrlPreview" style="display: none;">
                             <a id="eventUrl" class="fragment" href="#" target="_blank">
-                                <img src="@CfpExchange.Helpers.Constants.NoEventImageUrl" style="max-width: 116px;overflow: hidden;" alt="Event image" />
+                                <img src="@CfpExchange.Common.Constants.NoEventImageUrl" style="max-width: 116px;overflow: hidden;" alt="Event image" />
                                 <h3></h3>
                                 <p class="text"></p>
                             </a>
