+++ CfpExchange/appsettings.Development.json
@@ -19,6 +19,6 @@
   "FeatureToggle": {
     "HostOwnImages": true
   },
-  "ServicebusConnectionString": "FakeSBCS",
+  "ServicebusConnectionString": "Endpoint=sb://FakeEndpoint.servicebus.windows.net;SharedAccessKeyName=FakeSender;SharedAccessKey=FakeKey;EntityPath=FakeQueue;",
   "GooglePlacesApiKey": ""
 }
